% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_replace.R
\name{cat_replace}
\alias{cat_replace}
\title{FUNCTION_TITLE}
\usage{
cat_replace(
  b_v,
  g_v,
  threshold = NA,
  method = "jw",
  q = 1,
  p = 0,
  bt = 0,
  useBytes = FALSE,
  weight = c(d = 1, i = 1, t = 1)
)
}
\arguments{
\item{b_v}{PARAM_DESCRIPTION}

\item{g_v}{PARAM_DESCRIPTION}

\item{threshold}{PARAM_DESCRIPTION, Default: NA}

\item{method}{PARAM_DESCRIPTION, Default: 'jw'}

\item{q}{PARAM_DESCRIPTION, Default: 1}

\item{p}{PARAM_DESCRIPTION, Default: 0}

\item{bt}{PARAM_DESCRIPTION, Default: 0}

\item{useBytes}{PARAM_DESCRIPTION, Default: FALSE}

\item{weight}{PARAM_DESCRIPTION, Default: c(d = 1, i = 1, t = 1)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
