% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{clean_caterpillars}
\alias{clean_caterpillars}
\title{clean_caterpillars}
\format{
A data frame with 74 rows and 3 variables:
\describe{
\item{\code{species}}{character Full latin names of 29 caterpillar species.}
\item{\code{count}}{integer Randomly generated fake counts of the caterpillars.}
\item{\code{year}}{double Year of caterpillar observations.}
}
}
\usage{
clean_caterpillars
}
\description{
Dataframe with caterpillar counts from three summers.
}
\details{
An example dataset with clean caterpillar species names.
}
\keyword{datasets}
