% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{clean_names.df}
\alias{clean_names.df}
\title{clean_names.df}
\format{
A data frame with 20 rows and 2 variables:
\describe{
\item{\code{first}}{character Clean first names}
\item{\code{last}}{character Clean last names}
}
}
\usage{
clean_names.df
}
\description{
Data set of clean names
}
\details{
An example data that can be used in testing messy.cats functions
}
\keyword{datasets}
