% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{country.names}
\alias{country.names}
\title{country.names}
\format{
A data frame with 203 rows and 1 variables:
\describe{
\item{\code{name}}{character Names of countries}
}
}
\usage{
country.names
}
\description{
Dataframe with country names as only variable, contains many popular and official names for countries.
}
\details{
This dataframe contains a list of clean country names with many popular and official names for countries.
}
\keyword{datasets}
