% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_match.R
\name{country_match}
\alias{country_match}
\title{country_match}
\usage{
country_match(
  messy_countries,
  return_dists = FALSE,
  return_lists = NA,
  threshold = NA,
  pick_lists = F,
  p = 0
)
}
\arguments{
\item{messy_countries}{PARAM_DESCRIPTION}

\item{return_dists}{PARAM_DESCRIPTION, Default: FALSE}

\item{return_lists}{PARAM_DESCRIPTION, Default: NA}

\item{threshold}{PARAM_DESCRIPTION, Default: NA}

\item{pick_lists}{PARAM_DESCRIPTION, Default: F}

\item{p}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
