% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_replace.R
\name{country_replace}
\alias{country_replace}
\title{country_replace}
\usage{
country_replace(
  messy_countries,
  threshold = NA,
  method = "jw",
  q = 1,
  p = 0,
  bt = 0,
  useBytes = FALSE,
  weight = c(d = 1, i = 1, t = 1)
)
}
\arguments{
\item{messy_countries}{PARAM_DESCRIPTION}

\item{threshold}{PARAM_DESCRIPTION, Default: NA}

\item{method}{PARAM_DESCRIPTION, Default: 'jw'}

\item{q}{PARAM_DESCRIPTION, Default: 1}

\item{p}{PARAM_DESCRIPTION, Default: 0}

\item{bt}{PARAM_DESCRIPTION, Default: 0}

\item{useBytes}{PARAM_DESCRIPTION, Default: FALSE}

\item{weight}{PARAM_DESCRIPTION, Default: c(d = 1, i = 1, t = 1)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 lst <- c("Conagoa", "Blearaus", "Venzesual", "Uruagsya", "England")
 fixed <- country_replace(lst)
 }
}
}
