% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deerpop.R
\docType{data}
\name{deerpop}
\alias{deerpop}
\title{DATASET_TITLE}
\format{
A data frame with 324 rows and 6 variables:
\describe{
\item{\code{DMU}}{character COLUMN_DESCRIPTION}
\item{\code{Year}}{double COLUMN_DESCRIPTION}
\item{\code{AntleredHarvest}}{double COLUMN_DESCRIPTION}
\item{\code{AntlerlessHarvest}}{double COLUMN_DESCRIPTION}
\item{\code{TotalHarvest}}{double COLUMN_DESCRIPTION}
\item{\code{PosthuntPopEst}}{double COLUMN_DESCRIPTION}
}
}
\usage{
deerpop
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
