% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmu.R
\docType{data}
\name{dmu}
\alias{dmu}
\title{DATASET_TITLE}
\format{
A data frame with 103 rows and 9 variables:
\describe{
\item{\code{OBJECTID}}{double COLUMN_DESCRIPTION}
\item{\code{FID_WM_DEE}}{double COLUMN_DESCRIPTION}
\item{\code{DEER_MANAG}}{character COLUMN_DESCRIPTION}
\item{\code{FID_WM_DMU}}{character COLUMN_DESCRIPTION}
\item{\code{DEER_MGMT_}}{character COLUMN_DESCRIPTION}
\item{\code{METRO_SUBU}}{character COLUMN_DESCRIPTION}
\item{\code{DEER_MGM_1}}{character COLUMN_DESCRIPTION}
\item{\code{SHAPEAREA}}{double COLUMN_DESCRIPTION}
\item{\code{SHAPELEN}}{double COLUMN_DESCRIPTION}
}
}
\usage{
dmu
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
