% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{messy_caterpillars}
\alias{messy_caterpillars}
\title{messy_caterpillars}
\format{
A data frame with 39 rows and 3 variables:
\describe{
\item{\code{CaterpillarSpecies}}{character COLUMN_DESCRIPTION}
\item{\code{Avg Weight (mg)}}{double COLUMN_DESCRIPTION}
\item{\code{Avg Length (cm)}}{double COLUMN_DESCRIPTION}
}
}
\usage{
messy_caterpillars
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
