% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{messy_names.df}
\alias{messy_names.df}
\title{messy_names.df}
\format{
A data frame with 80 rows and 2 variables:
\describe{
\item{\code{first}}{character Messy first names}
\item{\code{last}}{character MEssy last names}
}
}
\usage{
messy_names.df
}
\description{
Data set of messy names
}
\details{
An example data set of messy names that can be used in testing messy.cats functions.
}
\keyword{datasets}
