% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{picked_list}
\alias{picked_list}
\title{picked_list}
\format{
A data frame with 15 rows and 3 variables:
\describe{
\item{\code{picked_list}}{character Handpicked matches of cars.}
}
}
\usage{
picked_list
}
\description{
Handpicked matches from the datasets in intro.rmd.
}
\details{
An example dataset of matched car names.
}
\keyword{datasets}
