% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{state.name}
\alias{state.name}
\title{state.name}
\format{
A data frame with 50 rows and 1 variables:
\describe{
\item{\code{states}}{character State names}
}
}
\usage{
state.name
}
\description{
Testing data set of state names
}
\details{
Testing data set of state names
}
\keyword{datasets}
