% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_match.R
\name{state_match}
\alias{state_match}
\title{state_match}
\usage{
state_match(
  messy_states,
  return_dists = FALSE,
  return_lists = NA,
  threshold = NA,
  pick_lists = F,
  p = 0
)
}
\arguments{
\item{messy_states}{PARAM_DESCRIPTION}

\item{return_dists}{PARAM_DESCRIPTION, Default: FALSE}

\item{return_lists}{PARAM_DESCRIPTION, Default: NA}

\item{threshold}{PARAM_DESCRIPTION, Default: NA}

\item{pick_lists}{PARAM_DESCRIPTION, Default: F}

\item{p}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
state_match() is a wrapper for cat_match() used when US states are misspelled.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
