% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sul05.R
\docType{data}
\name{sul05}
\alias{sul05}
\title{DATASET_TITLE}
\format{
A data frame with 140 rows and 6 variables:
\describe{
\item{\code{X1}}{double COLUMN_DESCRIPTION}
\item{\code{ISO3}}{character COLUMN_DESCRIPTION}
\item{\code{Year}}{double COLUMN_DESCRIPTION}
\item{\code{sulfer}}{double COLUMN_DESCRIPTION}
\item{\code{region}}{character COLUMN_DESCRIPTION}
\item{\code{logEmissions}}{double COLUMN_DESCRIPTION}
}
}
\usage{
sul05
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
