% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typos.R
\docType{data}
\name{typos}
\alias{typos}
\title{typos}
\format{
A data frame with 27 rows and 2 variables:
\describe{
\item{\code{occurrence}}{double number of times word appears in text}
\item{\code{species}}{character words in text}
}
}
\usage{
typos
}
\description{
Data set of words, some correctly spelled, some typos, with their occurrence in text
}
\details{
An example data set that can be used in testing fix_typos().
}
\keyword{datasets}
