% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wld.R
\docType{data}
\name{wld}
\alias{wld}
\title{DATASET_TITLE}
\format{
A data frame with 99338 rows and 7 variables:
\describe{
\item{\code{X1}}{double COLUMN_DESCRIPTION}
\item{\code{long}}{double COLUMN_DESCRIPTION}
\item{\code{lat}}{double COLUMN_DESCRIPTION}
\item{\code{group}}{double COLUMN_DESCRIPTION}
\item{\code{order}}{double COLUMN_DESCRIPTION}
\item{\code{region}}{character COLUMN_DESCRIPTION}
\item{\code{subregion}}{character COLUMN_DESCRIPTION}
}
}
\usage{
wld
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
