% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_match.R
\name{cat_match}
\alias{cat_match}
\title{cat_match}
\usage{
cat_match(
  b_v,
  g_v,
  return_dists = FALSE,
  return_lists = NA,
  pick_lists = F,
  threshold = NA,
  method = "jw",
  q = 1,
  p = 0,
  bt = 0,
  useBytes = FALSE
)
}
\arguments{
\item{b_v}{PARAM_DESCRIPTION}

\item{g_v}{PARAM_DESCRIPTION}

\item{return_dists}{PARAM_DESCRIPTION, Default: FALSE}

\item{return_lists}{PARAM_DESCRIPTION, Default: NA}

\item{pick_lists}{PARAM_DESCRIPTION, Default: F}

\item{threshold}{PARAM_DESCRIPTION, Default: NA}

\item{method}{PARAM_DESCRIPTION, Default: 'jw'}

\item{q}{PARAM_DESCRIPTION, Default: 1}

\item{p}{PARAM_DESCRIPTION, Default: 0}

\item{bt}{PARAM_DESCRIPTION, Default: 0}

\item{useBytes}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
cat_match(bad$DEER_MGM_1, good$DMU, p = 0.1, return_dists = T, return_lists = 3, pick_lists = F)
 }
}
}
