% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_match.R
\name{cat_match}
\alias{cat_match}
\title{cat_match}
\usage{
cat_match(
  messy_v,
  clean_v,
  return_dists = FALSE,
  return_lists = NA,
  pick_lists = F,
  threshold = NA,
  method = "jw",
  q = 1,
  p = 0,
  bt = 0,
  useBytes = FALSE,
  weight = c(d = 1, i = 1, t = 1)
)
}
\arguments{
\item{messy_v}{messy_vector: the vector of messy categories or strings}

\item{clean_v}{clean_vector: the vector of clean categories or strings to compare with messy_v}

\item{return_dists}{Returns the distance between the matched values. Range of values depends on method used: see help for stringdist::stringdist, Default: FALSE}

\item{return_lists}{Instead of the closest match, return a list of the top number of matches that you specify with this argument, Default: NA}

\item{pick_lists}{If TRUE cat_match() asks for user input and allows you to choose between the list of matches created by return_lists, Default: F}

\item{threshold}{Sets a threshold for an acceptable match. If return_lists != NA then lists are only returned for matches above the threshold, Default: NA}

\item{method}{method to calculate string distances. See help for stringdist::stringdist, Default: 'jw'}

\item{q}{Size of the q-gram; must be nonnegative. Only applies to method='qgram', 'jaccard' or 'cosine'., Default: 1}

\item{p}{Penalty factor for Jaro-Winkler distance. The valid range for p is 0 <= p <= 0.25. If p=0 (default), the Jaro-distance is returned. Applies only to method='jw', Default: 0}

\item{useBytes}{If TRUE, the matching is done byte-by-byte rather than character-by-character, Default: FALSE}

\item{weight}{For method='osa' or 'dl', the penalty for deletion, insertion, substitution and transposition, in that order. When method='lv', the penalty for transposition is ignored. When method='jw', the weights associated with characters of a, characters from b and the transposition weight, in that order. Weights must be positive and not exceed 1. weight is ignored completely when method='hamming', 'qgram', 'cosine', 'Jaccard', 'lcs', or soundex., Default: c(d = 1, i = 1, s = 1, t = 1)}

\item{btWinkler's}{boost threshold. Winkler's penalty factor is only applied when the Jaro distance is larger than bt. Applies only to method='jw' and p>0., Default: 0, Default: 0}
}
\value{
Returns a dataframe with each unique value in the bad vector and it's closest match in the good vector. If return_dists is TRUE the distances between the matches are added as a column.
}
\description{
cat_match() uses fuzzy matching to clean messy categories or strings of a bad vector by comparing them to a good vector.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
cat_match(bad$DEER_MGM_1, good$DMU, p = 0.1, return_dists = T, return_lists = 3, pick_lists = F)
 }
}
}
