% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_match.R
\name{cat_match}
\alias{cat_match}
\title{cat_match}
\usage{
cat_match(
  b_v,
  g_v,
  return_dists = FALSE,
  return_lists = NA,
  pick_lists = F,
  threshold = NA,
  method = "jw",
  q = 1,
  p = 0,
  bt = 0,
  useBytes = FALSE,
  weight = c(d = 1, i = 1, t = 1)
)
}
\arguments{
\item{b_v}{bad_vector: the vector of messy categories or strings}

\item{g_v}{good_vector: the vector of clean categories or strings to compare with b_V}

\item{return_dists}{and 1 means they share no characters, Default: FALSE}

\item{return_lists}{specify, Default: NA}

\item{pick_lists}{PARAM_DESCRIPTION, Default: F}

\item{threshold}{PARAM_DESCRIPTION, Default: NA}

\item{method}{method to calculate string distances. See help for stringdist::stringdist, Default: 'jw'}

\item{q}{PARAM_DESCRIPTION, Default: 1}

\item{p}{PARAM_DESCRIPTION, Default: 0}

\item{bt}{PARAM_DESCRIPTION, Default: 0}

\item{useBytes}{PARAM_DESCRIPTION, Default: FALSE}

\item{weight}{For method='osa', Default: c(d=1, i=1, t=1)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
cat_match() uses fuzzy matching to clean messy categories or strings of a bad vector by comparing them to a good vector. Returns a dataframe with each unique value in the bad vector and it's closest match in the good vector.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
cat_match(bad$DEER_MGM_1, good$DMU, p = 0.1, return_dists = T, return_lists = 3, pick_lists = F)
 }
}
}
