% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{clean_caterpillars}
\alias{clean_caterpillars}
\title{clean_caterpillars}
\format{
A data frame with 74 rows and 3 variables:
\describe{
\item{\code{species}}{character COLUMN_DESCRIPTION}
\item{\code{count}}{integer COLUMN_DESCRIPTION}
\item{\code{year}}{double COLUMN_DESCRIPTION}
}
}
\usage{
clean_caterpillars
}
\description{
Dataframe with caterpillar counts from three summers.
}
\details{
DETAILS
}
\keyword{datasets}
