% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_match.R
\name{country_match}
\alias{country_match}
\title{country_match}
\usage{
country_match(messy_countries, threshold = NA, p = 0)
}
\arguments{
\item{messy_countries}{Vector containing the messy country names that will be replaced
by the closest match from \code{country.names}}

\item{threshold}{The maximum distance that will form a match. If this argument
is specified, any element in the messy vector that has no match closer than
the threshold distance will be replaced with NA. Default: NA}

\item{p}{Only used with method "jw", the Jaro-Winkler penatly size. Default: 0}
}
\value{
\code{country_match()} returns a cleaned version of the bad vector, with each
element replaced by the most similar element of the good vector.
}
\description{
A wrapper function for \code{cat_match()}hat only requires an inputted
vector of messy countries. \code{country_match()} uses a built in clean list of
country names \code{country.names} as the reference clean vector.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 lst <- c("Conagoa", "Blearaus", "Venzesual", "Uruagsya", "England")
 matched <- country_match(lst)
 }
}
}
