% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messy.cats-package.R
\docType{package}
\name{messy.cats-package}
\alias{messy.cats}
\alias{messy.cats-package}
\title{messy.cats: Employs String Distance Tools to Help Clean Categorical Data}
\description{
Matching with string distance has never been easier! `messy.cats` contains various functions that employ string distance tools in order to make data management easier for users working with categorical data. Categorical data, especially user inputted categorical data that often tends to be plagued by typos, can be difficult to work with. `messy.cats` aims to provide functions that make cleaning categorical data simple and easy.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/hkarp1/messy.cats/issues}
}

}
\author{
\strong{Maintainer}: Harrison Karp \email{hkarp@wesleyan.edu}

Authors:
\itemize{
  \item Andrew Hennessy \email{ahennessy@wesleyan.edu}
  \item Duncan Kreutter \email{dkreutter@wesleyan.edu}
}

}
\keyword{internal}
