% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messy.cats-package.R
\docType{package}
\name{messy.cats-package}
\alias{messy.cats}
\alias{messy.cats-package}
\title{messy.cats: Employs String Distance Tools to Help Clean Categorical Data}
\description{
Messy.cats contains various functions that employ string distance tools in order to make data management easier for users working with categorical data. Categorical data, especially user inputted categorical data that often tends to be plagued by typos, can be difficult to work with. Messy.cats aims to provide functions that make cleaning categorical data simple and easy.
}
\author{
\strong{Maintainer}: Duncan Kreutter \email{dkreutter@wesleyan.edu}

}
\keyword{internal}
