% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{messy_caterpillars}
\alias{messy_caterpillars}
\title{messy_caterpillars}
\format{
A data frame with 39 rows and 3 variables:
\describe{
\item{\code{CaterpillarSpecies}}{character Full latin names of 39 caterpillar
species with spelling and formatting errors.}
\item{\code{Avg Weight (mg)}}{double Randomly generated fake weight data for each caterpillar species.}
\item{\code{Avg Length (cm)}}{double Randomly generated fake length data for each caterpillar species.}
}
}
\usage{
messy_caterpillars
}
\description{
DATASET_DESCRIPTION
}
\details{
An example dataset with messy caterpillar species names.
}
\keyword{datasets}
