% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{messy_states1}
\alias{messy_states1}
\title{messy_states1}
\format{
A data frame with 50 rows and 1 variables:
\describe{
\item{\code{messy_states1}}{character All 50 US states with 1 randomly changed character.}
}
}
\usage{
messy_states1
}
\description{
US State names with 1 character randomly changed.
}
\details{
An example dataset with mispelled US state names. The names have had 1 character randomly changed.
}
\keyword{datasets}
