% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{messy_states2}
\alias{messy_states2}
\title{messy_states2}
\format{
A data frame with 50 rows and 1 variables:
\describe{
\item{\code{messy_states2}}{character All 50 US states with 2 randomly changed characters.}
}
}
\usage{
messy_states2
}
\description{
US State names with 2 characters randomly changed.
}
\details{
An example dataset with mispelled US state names. The names have had 2 characters randomly changed.
}
\keyword{datasets}
