% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piced_list.R
\docType{data}
\name{picked_list}
\alias{picked_list}
\title{picked_list}
\format{
A data frame with 15 rows and 3 variables:
\describe{
\item{\code{bad}}{character column of bad car names}
\item{\code{match}}{character column of good car names}
\item{\code{dists}}{double string distance between the good and bad car names}
}
}
\usage{
picked_list
}
\description{
Handpicked matches from the datasets in intro.rmd.
}
\details{
An example dataset of matched car names.
}
\keyword{datasets}
