% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_replace.R
\name{state_replace}
\alias{state_replace}
\title{state_replace}
\usage{
state_replace(
  messy_states,
  threshold = NA,
  method = "jw",
  q = 1,
  p = 0,
  bt = 0,
  useBytes = FALSE,
  weight = c(d = 1, i = 1, t = 1)
)
}
\arguments{
\item{messy_states}{PARAM_DESCRIPTION}

\item{threshold}{PARAM_DESCRIPTION, Default: NA}

\item{method}{PARAM_DESCRIPTION, Default: 'jw'}

\item{q}{PARAM_DESCRIPTION, Default: 1}

\item{p}{PARAM_DESCRIPTION, Default: 0}

\item{bt}{PARAM_DESCRIPTION, Default: 0}

\item{useBytes}{PARAM_DESCRIPTION, Default: FALSE}

\item{weight}{PARAM_DESCRIPTION, Default: c(d = 1, i = 1, t = 1)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Wrapper of cat_replace() Calls cat_replace on input with clean vector state.name, a vector of the names of all U.S. states
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
